main_project/requirements.txt
asgiref==3.8.1
Django==4.2.14
djangorestframework==3.15.2
sqlparse==0.5.0
typing_extensions==4.12.2
psycopg2-binary==2.9.9